Split stack traits into separate modules for TCP and UDP
